Major changes made to the "generateDriver" function of the kernel. It works as follows: For each value parameter 'v' in the IR node case class, do:    1. Declare 'v' with appropriate type, as given by remap.    2. if 'v' is a primitive type, assume integer, and use atoi on argv to    read the argument    3. If 'v' is not a primitive type, assume existence of a kernel named    'kernel_v'. Generate the statement 'v = kernel_v'.
Generate the kernel call after all the kernel_* functions to initialize
arguments have been called.

TODO: The arguments to each of the 'kernel_v' functions have to yet be
resolved. To be precise, from the main kernel's IR node, we need a way
to figure out:
1. How many arguments are to be read from command line, and their types
2. How these arguments read are used in the 'kernel_v' calls to
initialize the data structures passed into the kernel