More changes to generation of driver - simple genetic search with one generation works. Problem exists in stitching the right values back into the source code, though.
One hypothesis - I have not added the cases to 'mirror' the new IR
nodes, maybe that is causing the issue. Hence, I have created a separate
node 'Densematrix_matmult_autotune3' which is the same as the vanilla
'Densematrix_matmult' except that it does not have mirroring code. Test,
find out